Create quote model
This is where we'll store everything Rails needs to know about the quote
model. For now we don't need to tell it anything other than the fact it
exists, which is what an empty model will do. Rails will use the corresponding
Quotes table in the database based on the name of this model.