Add next quote route, controller action and link
This adds a route so we can do a get request back to our quotes controller.
We've directed this to the 'next' action in the controller, so Rails will run
the code in that action.

In this case it will fetch another random quote and render the index again
using JavaScript only with the new variable we've assigned, which should be a
new quote.

This doesn't handle the case where the next quote chosen is the same as the
first, which is a bug that some users will encounter.